+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/util/VariablesService.java
@@ -161,6 +161,21 @@ public class VariablesService {
      */
     private String daysOfSupplySubmodelAssetId;
 
+    /**
+     * The url under which this application's anonymized planned production request endpoint can
+     * be reached by external machines.
+     */
+    public String getProductionAnonymizedSubmodelEndpoint() {
+        return getPurisBaseUrl() + getContextPath() + "planned-production-output/anonymized/request";
+    }
+
+    @Value("${puris.productionanonymizedsubmodel.apiassetid}")
+    /**
+     * The assetId that shall be assigned to the Anonymized Planned Production request API
+     * during asset creation.
+     */
+    private String productionAnonymizedSubmodelApiAssetId;
+
     @Value("${puris.frameworkagreement.credential}")
     /**
      * The name of the framework agreement to be used.
@@ -314,6 +329,10 @@ public class VariablesService {
 
     public String getDaysOfSupplySubmodelApiAssetId() {
         return daysOfSupplySubmodelAssetId + "@" + ownBpnl;
+    }
+
+    public String getProductionAnonymizedSubmodelApiAssetId() {
+        return productionAnonymizedSubmodelApiAssetId + "@" + ownBpnl;
     }
 
     public String getNotificationApiAssetId() {